pyproject.toml
[project]
name = "proxycraft"
version = "1.0.0"
description = "⚡ High-performance Python proxy server with authentication, load balancing, caching, and multi-protocol support"
authors = [{name = "Sylvain Mouquet", email = "sylvain.mouquet@gmail.com"}]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofile>=3.9.0",
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.18",
    "antpathmatcher>=1.0.0",
    "apscheduler>=3.11.0",
    "curl-cffi>=0.11.2",
    "granian>=2.3.3",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "hypercorn>=0.17.3",
    "pydantic>=2.11.4",
    "python-jose>=3.4.0",
    "robyn>=0.68.0",
    "starlette>=0.50",
    "structlog>=25.3.0",
    "uvicorn>=0.40",
    "uvloop>=0.21.0",
]

license = { text = "MIT" }
url = "https://github.com/sylvainmouquet/proxycraft"
keywords = [
    "proxycraft",
    "proxy",
    "python",
    "http",
    "https",
    "websocket",
    "socks5",
    "networking",
    "reverse-proxy",
    "authentication",
    "load-balancing"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
homepage = "https://github.com/sylvainmouquet/proxycraft"
documentation = "https://github.com/sylvainmouquet/proxycraft"
repository = "https://github.com/sylvainmouquet/proxycraft"
changelog = "https://github.com/sylvainmouquet/proxycraft/releases"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
  "certifi>=2025.4.26",
  "pip>=24.2",
  "pytest-asyncio>=0.24.0",
  "pytest-mock>=3.14.0",
  "pytest>=8.3.3",
  "ruff>=0.6.7"
]

[tool.uv.sources]
proxycraft = { path = "proxycraft" }

[project.entry-points.pytest11]
proxycraft = "proxycraft"
